requirements.txt
alembic==1.2.1
aniso8601==8.0.0
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
flask-marshmallow==0.10.1
Flask-Migrate==2.3.0
Flask-MySQLdb==0.2.0
Flask-RESTful==0.3.7
Flask-Script==2.0.6
Flask-SQLAlchemy==2.1
google-api-core==1.14.3
google-auth==1.6.3
google-cloud-core==1.0.3
google-cloud-storage==1.21.0
google-resumable-media==0.4.1
googleapis-common-protos==1.6.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==2.19.2
marshmallow-sqlalchemy==0.16.1
mysqlclient==1.4.4
pkg-resources==0.0.0
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
python-dateutil==2.8.0
python-editor==1.0.4
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
rsa==4.0
six==1.12.0
SQLAlchemy==1.3.10
urllib3==1.25.6
Werkzeug==0.16.0
